dom/battery/test/test_battery_charging.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Battery API</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

"use strict";

SimpleTest.waitForExplicitFinish();

/** Test for Battery API **/
navigator.getBattery().then(function (battery) {
  ok(battery.level >= 0.0 && battery.level <= 1.0, "Battery level " + battery.level + " should be in the range [0.0, 1.0]");

  SpecialPowers.pushPrefEnv({"set": [["dom.battery.test.charging", true]]}, function () {
    is(battery.charging, true, "Battery should be charging");
    ok(battery.chargingTime >= 0, "Battery chargingTime " + battery.chargingTime + " should be nonnegative when charging");
    is(battery.dischargingTime, Infinity, "Battery dischargingTime should be Infinity when charging");

    SimpleTest.finish();
  });
});

</script>
</pre>
</body>
</html>
